Updated Null checks

+++ Reuse Book/src/main/java/com/example/reusebook/Service/AuthorService.java
@@ -1,6 +1,7 @@
 package com.example.reusebook.Service;
 
 import com.example.reusebook.Model.Author;
+import com.example.reusebook.Model.Book;
 import com.example.reusebook.Repository.AuthorRepository;
 import com.example.reusebook.Repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +11,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class AuthorService {
@@ -18,12 +20,20 @@ public class AuthorService {
     @Autowired
     private AuthorRepository authorRepository;
 
-    public ResponseEntity<Author> addAuthor(Long bookId, @RequestBody Author author){
-        Author a = bookRepository.findById(bookId).map(book -> {
-            author.setBook(book);
-            return authorRepository.save(author);
-        }).orElseThrow();
-        return new ResponseEntity<>(a, HttpStatus.CREATED);
+    public ResponseEntity<Object> addAuthor(Long bookId, @RequestBody Author author){
+        Optional<Book> book = bookRepository.findById(bookId);
+        if(book.isPresent()){
+            Book b = book.get();
+            if(author.getName() != null && !author.getName().isBlank() && !author.getName().isEmpty()){
+                author.setBook(b);
+                return new ResponseEntity<>(authorRepository.save(author),HttpStatus.CREATED);
+            }else{
+                return new ResponseEntity<>("Bad Request for Author", HttpStatus.BAD_REQUEST);
+            }
+
+        }else{
+            return new ResponseEntity<>("Book Id not found", HttpStatus.BAD_REQUEST);
+        }
     }
 
     public ResponseEntity<List<Author>> getAllAuthorByBookId(Long bookId) {
@@ -31,13 +41,25 @@ public class AuthorService {
         return new ResponseEntity<>(authors, HttpStatus.OK);
     }
 
-    public ResponseEntity<Author> updateAuthor(long authorId,Author authorR) {
-        Author author = authorRepository.findById(authorId).orElseThrow();
-        author.setName(authorR.getName());
-        return new ResponseEntity<>(authorRepository.save(author), HttpStatus.OK);
+    public ResponseEntity<Object> updateAuthor(long authorId,Author authorR) {
+        Optional<Author> author = authorRepository.findById(authorId);
+        if(author.isPresent()){
+            Author a = author.get();
+            if(authorR.getName() != null && !authorR.getName().isBlank() && !authorR.getName().isEmpty()) {
+                a.setName(authorR.getName());
+                return new ResponseEntity<>(authorRepository.save(a), HttpStatus.OK);
+            }else{
+                return new ResponseEntity<>("Bad Request for Author", HttpStatus.BAD_REQUEST);
+            }
+        }else {
+            return new ResponseEntity<>("Bad Request for Author", HttpStatus.BAD_REQUEST);
+        }
     }
 
     public ResponseEntity<HttpStatus> deleteAuthor(long id) {
+        Optional<Author> author = authorRepository.findById(id);
+        if(!author.isPresent())
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         authorRepository.deleteById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
